requirements.txt
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
drf-spectacular==0.27.0
mysqlclient==2.2.0
Pillow==10.0.1
celery==5.3.6
redis==5.0.1